Move test to different file.
